extract form component
